+++ web-console/src/react-table/index.ts
@@ -16,6 +16,6 @@
  * limitations under the License.
  */
 
+export * from './react-table-filters';
 export * from './react-table-inputs';
 export * from './react-table-pagination/react-table-pagination';
-export * from './react-table-utils';
